[plugin] Use imp module in order to retrieve info
This commit provides an alternative to the "cat/grep" method for retrieving the
information of a plugin.
Instead, it uses the imp module to load each module and then retrieve the
information directly from the plugin (like the description).

Plus, it prepares the merge for the issue #37 by providing a boolean describing
whether the plugin uses internets resources or not.