[plugin] Define an 'attr' dict for each plugin.
The 'ATTR' disctionary gathers the attributes of a plugin. Each new variable
will have to be declared into this dictionary.
For instance, the USE_INTERNET_RESOURCES becomes the 'INTERNET_RESOURCES' into
the ATTR dict.

Moreover, the ATTR dict is converted into json in order to be saved into the
database, in the attr column.