[report] Update display when the user selects a ranking for a plugin.
+ With some javascript, the tab concerning a plugin will be colorized depending
on the ranking selected by the user.
+ Instead of having a gray color when the ranking comes from OWTF, the opacity
is set to 0.5 instead.

FIXME:
A bug is known but I can't figure out a way to fix it.

When the page is reloaded, the colors of each tab are lost. They only reapear
when the user selects again the ranking value.

A fix would be to automatically check the ranking when the page is loaded and
ready and change the class of the tab according to that value.