Reorder the initialisation sequence, Fixes #367
The general config file is loaded by 'db_config' component. It contains all the necessary data required by other components. This PR shifts the loading of the config files to phase2 so that processed arguments are available. Due to shift of 'db_config' initialization to phase2, all the components dependent on it are shifted in phase 2 (Like: 'timer'). The components shifted to phase2 are initialized in order as before. The initialization is divided in two part to deal with circular dependency.